# docs/swagger.yaml
definitions:
  db.CardModel:
    properties:
      authToken:
        type: string
      cardType:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      expiryMonth:
        type: string
      expiryYear:
        type: string
      id:
        type: string
      isApproved:
        type: boolean
      isDefault:
        type: boolean
      lastDigits:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - authToken
    - cardType
    - createdAt
    - deletedAt
    - expiryMonth
    - expiryYear
    - id
    - isApproved
    - isDefault
    - lastDigits
    - updatedAt
    - user
    - userId
    type: object
  db.Currency:
    enum:
    - NGN
    type: string
    x-enum-varnames:
    - CurrencyNgn
  db.DeploymentLogModel:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      deployment:
        $ref: '#/definitions/db.DeploymentModel'
      deploymentId:
        type: string
      id:
        type: string
      jobId:
        type: string
      message:
        type: string
      updatedAt:
        type: string
    required:
    - createdAt
    - deletedAt
    - deployment
    - deploymentId
    - id
    - jobId
    - message
    - updatedAt
    type: object
  db.DeploymentModel:
    properties:
      DeploymentLog:
        items:
          $ref: '#/definitions/db.DeploymentLogModel'
        type: array
      actor:
        type: string
      commitHash:
        type: string
      commitMessage:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      machine:
        $ref: '#/definitions/db.MachineModel'
      machineId:
        type: string
      repoConnection:
        $ref: '#/definitions/db.RepoConnectionModel'
      repoConnectionId:
        type: string
      status:
        $ref: '#/definitions/db.DeploymentStatus'
      updatedAt:
        type: string
    required:
    - DeploymentLog
    - actor
    - commitHash
    - commitMessage
    - createdAt
    - deletedAt
    - id
    - machine
    - machineId
    - repoConnection
    - repoConnectionId
    - status
    - updatedAt
    type: object
  db.DeploymentStatus:
    enum:
    - IDLE
    - IN_PROGRESS
    - SUCCESSFUL
    - FAILED
    type: string
    x-enum-varnames:
    - DeploymentStatusIdle
    - DeploymentStatusInProgress
    - DeploymentStatusSuccessful
    - DeploymentStatusFailed
  db.GithubAccountConnectionModel:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      githubEmail:
        type: string
      githubId:
        type: string
      githubInstallationId:
        type: integer
      githubUsername:
        type: string
      id:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - createdAt
    - deletedAt
    - githubEmail
    - githubId
    - githubInstallationId
    - githubUsername
    - id
    - updatedAt
    - user
    - userId
    type: object
  db.InvoiceModel:
    properties:
      createdAt:
        type: string
      currency:
        $ref: '#/definitions/db.Currency'
      deletedAt:
        type: string
      description:
        type: string
      from:
        type: string
      id:
        type: string
      productId:
        type: string
      quantity:
        type: integer
      reference:
        type: string
      status:
        $ref: '#/definitions/db.InvoiceStatus'
      to:
        type: string
      totalPrice:
        type: number
      unitPrice:
        type: number
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - createdAt
    - currency
    - deletedAt
    - description
    - from
    - id
    - productId
    - quantity
    - reference
    - status
    - to
    - totalPrice
    - unitPrice
    - updatedAt
    - user
    - userId
    type: object
  db.InvoiceStatus:
    enum:
    - PAID
    - UNPAID
    type: string
    x-enum-varnames:
    - InvoiceStatusPaid
    - InvoiceStatusUnpaid
  db.MachineModel:
    properties:
      Deployment:
        items:
          $ref: '#/definitions/db.DeploymentModel'
        type: array
      Network:
        items:
          $ref: '#/definitions/db.NetworkModel'
        type: array
      RepoConnection:
        items:
          $ref: '#/definitions/db.RepoConnectionModel'
        type: array
      containerId:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      machineImage:
        type: string
      machineName:
        type: string
      machineStatus:
        $ref: '#/definitions/db.MachineStatus'
      owner:
        $ref: '#/definitions/db.UserModel'
      ownerId:
        type: string
      plan:
        $ref: '#/definitions/db.MachinePlanModel'
      planId:
        type: string
      updatedAt:
        type: string
    required:
    - Deployment
    - Network
    - RepoConnection
    - containerId
    - createdAt
    - deletedAt
    - id
    - machineImage
    - machineName
    - machineStatus
    - owner
    - ownerId
    - plan
    - planId
    - updatedAt
    type: object
  db.MachinePlanModel:
    properties:
      Machine:
        items:
          $ref: '#/definitions/db.MachineModel'
        type: array
      cpu:
        type: string
      createdAt:
        type: string
      currency:
        $ref: '#/definitions/db.Currency'
      deletedAt:
        type: string
      hourlyRate:
        type: number
      id:
        type: string
      memory:
        type: string
      monthlyRate:
        type: number
      name:
        type: string
      updatedAt:
        type: string
    required:
    - Machine
    - cpu
    - createdAt
    - currency
    - deletedAt
    - hourlyRate
    - id
    - memory
    - monthlyRate
    - name
    - updatedAt
    type: object
  db.MachineStatus:
    enum:
    - CREATING
    - RUNNING
    - REBOOTING
    - SHUTTING_DOWN
    - SHUTDOWN
    type: string
    x-enum-varnames:
    - MachineStatusCreating
    - MachineStatusRunning
    - MachineStatusRebooting
    - MachineStatusShuttingDown
    - MachineStatusShutdown
  db.NetworkModel:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      destinationPort:
        type: integer
      hostName:
        type: string
      id:
        type: string
      ingressId:
        type: string
      listeningPort:
        type: integer
      machine:
        $ref: '#/definitions/db.MachineModel'
      machineId:
        type: string
      protocol:
        type: string
      serviceId:
        type: string
      updatedAt:
        type: string
    required:
    - createdAt
    - deletedAt
    - destinationPort
    - hostName
    - id
    - ingressId
    - listeningPort
    - machine
    - machineId
    - protocol
    - serviceId
    - updatedAt
    type: object
  db.RepoConnectionModel:
    properties:
      Deployment:
        items:
          $ref: '#/definitions/db.DeploymentModel'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      machine:
        $ref: '#/definitions/db.MachineModel'
      machineId:
        type: string
      repoId:
        type: string
      repoName:
        type: string
      updatedAt:
        type: string
    required:
    - Deployment
    - createdAt
    - deletedAt
    - id
    - machine
    - machineId
    - repoId
    - repoName
    - updatedAt
    type: object
  db.SocialConnectionModel:
    properties:
      connectionId:
        type: string
      connectionType:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - connectionId
    - connectionType
    - createdAt
    - deletedAt
    - id
    - updatedAt
    - user
    - userId
    type: object
  db.TransactionModel:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      currency:
        $ref: '#/definitions/db.Currency'
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: string
      reference:
        type: string
      status:
        $ref: '#/definitions/db.TransactionStatus'
      type:
        $ref: '#/definitions/db.TransactionType'
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    required:
    - amount
    - createdAt
    - currency
    - deletedAt
    - description
    - id
    - reference
    - status
    - type
    - updatedAt
    - user
    - userId
    type: object
  db.TransactionStatus:
    enum:
    - PENDING
    - SUCCESS
    - FAILED
    type: string
    x-enum-varnames:
    - TransactionStatusPending
    - TransactionStatusSuccess
    - TransactionStatusFailed
...
  db.UserModel:
    properties:
      Card:
        items:
          $ref: '#/definitions/db.CardModel'
        type: array
      GithubAccountConnection:
        items:
          $ref: '#/definitions/db.GithubAccountConnectionModel'
        type: array
      Invoice:
        items:
          $ref: '#/definitions/db.InvoiceModel'
        type: array
      Machine:
        items:
          $ref: '#/definitions/db.MachineModel'
        type: array
      SocialConnection:
        items:
          $ref: '#/definitions/db.SocialConnectionModel'
        type: array
      Transaction:
        items:
          $ref: '#/definitions/db.TransactionModel'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      password:
        type: string
      roles:
        items:
          type: string
        type: array
      updatedAt:
        type: string
    required:
    - Card
    - GithubAccountConnection
    - Invoice
    - Machine
    - SocialConnection
    - Transaction
    - createdAt
    - deletedAt
    - email
    - firstName
    - id
    - lastName
    - password
    - roles
    - updatedAt
    type: object
...
  types.CreateMachineArgs:
    properties:
      machineImage:
        type: string
      machineName:
        type: string
      planId:
        type: string
    required:
    - machineImage
    - machineName
    - planId
    type: object
  types.CreateMachinePlanArgs:
    properties:
      cpu:
        type: string
      currency:
        type: string
      memory:
        type: string
      monthlyRate:
        type: integer
      name:
        type: string
    required:
    - cpu
    - currency
    - memory
    - monthlyRate
    - name
    type: object
  types.CreateNetworkArgs:
    properties:
      destinationPort:
        type: integer
      listeningPort:
        type: integer
      machineId:
        type: string
      protocol:
        type: string
    required:
    - destinationPort
    - listeningPort
    - machineId
    - protocol
    type: object
  types.CreateRepoConnectionArgs:
    properties:
      machineId:
        type: string
      repoId:
        type: string
      repoName:
        type: string
    required:
    - machineId
    - repoId
    - repoName
    type: object
...
  types.UpdateMachineArgs:
    properties:
      machineImage:
        type: string
      machineName:
        type: string
      ownerId:
        type: string
      planId:
        type: string
    type: object
  types.UpdateMachinePlanArgs:
    properties:
      cpu:
        description: HourlyRate  *int32  `json:"hourlyRate" swag-validate:"optional"`
        type: string
      currency:
        type: string
      memory:
        type: string
      monthlyRate:
        type: integer
      name:
        type: string
    type: object
  types.UpdateRepoConnectionArgs:
    properties:
      machineId:
        type: string
      repoId:
        type: string
      repoName:
        type: string
    type: object
...
info:
  contact:
    email: formatio@overal-x.org
    name: Formatio Team
  title: Formatio API
  version: "1.0"
paths:
...
  /machine:
    get:
      consumes:
      - application/json
      operationId: listMachines
      parameters:
      - in: query
        name: search
        type: string
      - in: query
        name: skip
        type: integer
      - in: query
        name: sortBy
        type: string
      - in: query
        name: take
        type: integer
      - in: query
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.MachineModel'
            type: array
      tags:
      - machines
    post:
      consumes:
      - application/json
      operationId: createMachine
      parameters:
      - description: List Machine
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.CreateMachineArgs'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.MachineModel'
      tags:
      - machines
  /machine/{machineId}:
    delete:
      consumes:
      - application/json
      operationId: deleteMachine
      parameters:
      - description: Machine Id
        in: path
        name: machineId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      tags:
      - machines
    get:
      consumes:
      - application/json
      operationId: getMachine
      parameters:
      - description: Machine Id
        in: path
        name: machineId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.MachineModel'
      tags:
      - machines
    patch:
      consumes:
      - application/json
      operationId: updateMachine
      parameters:
      - description: Machine Id
        in: path
        name: machineId
        required: true
        type: string
      - description: Update Machine
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.UpdateMachineArgs'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
      tags:
      - machines
  /network:
    get:
      consumes:
      - application/json
      operationId: listNetworks
      parameters:
      - in: query
        name: machineId
        type: string
      - in: query
        name: search
        type: string
      - in: query
        name: skip
        type: integer
      - in: query
        name: sortBy
        type: string
      - in: query
        name: take
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.NetworkModel'
            type: array
      tags:
      - network
    post:
      consumes:
      - application/json
      operationId: createNetwork
      parameters:
      - description: List Networks
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.CreateNetworkArgs'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.NetworkModel'
      tags:
      - network
  /network/{networkId}:
    delete:
      consumes:
      - application/json
      operationId: deleteNetwork
      parameters:
      - description: Network Id
        in: path
        name: networkId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      tags:
      - network
  /plans:
    get:
      consumes:
      - application/json
      operationId: listMachinePlans
      parameters:
      - in: query
        name: search
        type: string
      - in: query
        name: skip
        type: integer
      - in: query
        name: sortBy
        type: string
      - in: query
        name: take
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.MachinePlanModel'
            type: array
      tags:
      - plans
    post:
      consumes:
      - application/json
      operationId: createMachinePlan
      parameters:
      - description: Create Machine Plan
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.CreateMachinePlanArgs'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.MachinePlanModel'
      tags:
      - plans
  /plans/{machinePlanId}:
    delete:
      consumes:
      - application/json
      operationId: deleteMachinePlan
      parameters:
      - description: Machine Plan Id
        in: path
        name: machinePlanId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      tags:
      - plans
    get:
      consumes:
      - application/json
      operationId: getMachinePlan
      parameters:
      - description: Machine Plan Id
        in: path
        name: machinePlanId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.MachinePlanModel'
      tags:
      - plans
    patch:
      consumes:
      - application/json
      operationId: updateMachinePlan
      parameters:
      - description: Machine Plan Id
        in: path
        name: machinePlanId
        required: true
        type: string
      - description: Update Machine
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.UpdateMachinePlanArgs'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/db.MachinePlanModel'
      tags:
      - plans
  /repo-connection:
    get:
      consumes:
      - application/json
      operationId: listRepoConnections
      parameters:
      - in: query
        name: machineId
        type: string
      - in: query
        name: search
        type: string
      - in: query
        name: skip
        type: integer
      - in: query
        name: sortBy
        type: string
      - in: query
        name: take
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.RepoConnectionModel'
            type: array
      tags:
      - repo-connection
    post:
      consumes:
      - application/json
      operationId: createRepoConnection
      parameters:
      - description: Create Repo Connection
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.CreateRepoConnectionArgs'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.RepoConnectionModel'
      tags:
      - repo-connection
  /repo-connection/{connectionId}:
    delete:
      consumes:
      - application/json
      operationId: deleteRepoConnection
      parameters:
      - description: Repo Connection Id
        in: path
        name: connectionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      tags:
      - repo-connection
    get:
      consumes:
      - application/json
      operationId: getRepoConnection
      parameters:
      - description: Repo Connection Id
        in: path
        name: connectionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/db.RepoConnectionModel'
      tags:
      - repo-connection
    patch:
      consumes:
      - application/json
      operationId: updateRepoConnection
      parameters:
      - description: Repo Connection Id
        in: path
        name: connectionId
        required: true
        type: string
      - description: Update Repo Connection
        in: body
        name: args
        required: true
        schema:
          $ref: '#/definitions/types.UpdateRepoConnectionArgs'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/db.RepoConnectionModel'
      tags:
      - repo-connection
...
swagger: "2.0"